backend/requirements.txt
fastapi
uvicorn
gunicorn
google-cloud-firestore
google-cloud-storage
google-generativeai
firebase_admin
pydantic[email]
python-multipart
PyPDF2
python-docx
google-cloud-tasks
